LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/function-invalidation-expected.txt


FAIL Appending a rule assert_equals: expected "10px" but got "42px"
FAIL Prepending a rule assert_equals: expected "10px" but got "42px"
FAIL Deleting a rule assert_equals: expected "10px" but got "42px"
FAIL Prepending a rule, then deleting last assert_equals: expected "10px" but got "42px"
